Fixes bugs at fork and clean instructions
Fixes bug in 'memfree'.
Clears name in 'cleaninstruction'.
Fixes signed vs unsigned integer comparisons.
Fixes bugs in 'forkProcess'.
Fixes bugs in 'cleanProgram'.
Adds name of process to 'pcbreport'.

+++ src/fcfs.c
@@ -24,7 +24,7 @@
 int fcfs(PCB *pcb, MEMORY *mem, int pcbindex) {
     /* Chegou ao fim da tabela PCB, não há mais processos em fila */
     debug("Working on PCB index %d.\n", pcbindex);
-    if (pcbindex >= pcb->top)
+    if ((size_t) pcbindex >= pcb->top)
         return FCFS_END;
     
     process *p = &(pcb->proc[pcbindex]);
